labs/hybrid-semantic-analysis/requirements.txt
# 核心依赖
pyyaml>=6.0
anthropic>=0.40.0
openai>=1.0.0

# 依存文法工具
ltp>=4.2.0
hanlp>=2.1.0

# 小模型推理 (Qwen/GLM)
transformers>=4.40.0
torch>=2.0.0
sentencepiece>=0.2.0
accelerate>=0.20.0

# 评估工具
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter (可选)
jupyter>=1.0.0
ipykernel>=6.0.0

# 文本处理
regex>=2023.0.0

# Token计数
tiktoken>=0.5.0

# 进度条
tqdm>=4.65.0
